Add mechanism to force link torchtrt_plugins lib on windows
if no symbol are used from the library, the library is pruned by the
linker. Here, we artificially use a dummy function in plugins.h to force
the linker to link to this library as soon as the header plugins.h is
included.

That mechanism is very similar to how it's done in pytorch/vision